moved full_node functions to full_node file
